improve device scheduling

+++ system/src/device.rs
@@ -4,7 +4,7 @@ pub struct Device {
     pub bus_msg_send: Option<(u8, u8)>,
     pub bus_msg_rcv: Option<(u8, u8)>,
     pub kernel: Box<dyn Kernel>,
-    pub clock_freq: u16,
+    pub clock_freq: u32,
     pub verbose: bool,
     pub addr: u8,
     pub ticks: u64,
@@ -12,7 +12,7 @@ pub struct Device {
 
 
 impl Device {
-    pub fn new(mut kernel: Box<dyn Kernel>, addr: u8, clock_freq: u16, verbose: bool) -> Self {
+    pub fn new(mut kernel: Box<dyn Kernel>, addr: u8, clock_freq: u32, verbose: bool) -> Self {
         kernel.init_bus(addr);
         
         Self {
